hw8: MvpPresenter

+++ app/src/main/java/ru/gorshenev/themesstyles/data/repositories/chat/ChatRepository.kt
@@ -2,112 +2,116 @@ package ru.gorshenev.themesstyles.data.repositories.chat
 
 import io.reactivex.Completable
 import io.reactivex.Observable
+import io.reactivex.Scheduler
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import ru.gorshenev.themesstyles.data.database.dao.MessageDao
-import ru.gorshenev.themesstyles.data.database.entities.MessageEntity
+import ru.gorshenev.themesstyles.data.database.entities.MessageWithReactionsEntity
 import ru.gorshenev.themesstyles.data.network.ZulipApi
 import ru.gorshenev.themesstyles.data.network.model.*
 import ru.gorshenev.themesstyles.data.repositories.chat.ChatMapper.toDomain
 import ru.gorshenev.themesstyles.data.repositories.chat.ChatMapper.toEntity
 import ru.gorshenev.themesstyles.domain.model.chat.MessageModel
 
-class ChatRepository(private val messageDao: MessageDao, private val api: ZulipApi) {
+class ChatRepository(
+    private val messageDao: MessageDao,
+    private val api: ZulipApi,
+    private val executionScheduler: Scheduler = Schedulers.io()
+) {
 
+    fun getMessage(messageId: Int) = api.getMessage(messageId).subscribeOn(executionScheduler)
 
-    fun getMessagesFromDb(topicName: String): Single<List<MessageModel>> {
-        return messageDao.getMessages(topicName)
-            .map { messageWithReactions -> messageWithReactions.toDomain() }
-            .subscribeOn(Schedulers.io())
+    fun getMessages(
+        streamName: String,
+        topicName: String,
+        anchorMessageId: Long,
+        numBefore: Int,
+        onlyRemote: Boolean
+    ): Observable<List<MessageModel>> {
+        val remoteMessages = getMessagesRemote(streamName, topicName, anchorMessageId, numBefore)
+        return if (onlyRemote) {
+            remoteMessages.map { it.messages.toDomain() }.toObservable()
+                .subscribeOn(executionScheduler)
+        } else {
+            Single.concatArrayEager(
+                getMessagesLocal(topicName).map { it.toDomain() },
+                remoteMessages.doOnSuccess { replaceLocalMessages(topicName, it.messages) }
+                    .map { it.messages.toDomain() }
+            ).toObservable().subscribeOn(executionScheduler)
+        }
     }
 
-    fun getMessagesFromApi(streamName: String, topicName: String): Single<GetMessageResponse> {
+    private fun replaceLocalMessages(topicName: String, remoteMessages: List<MessageResponse>) {
+        val messageEntities = remoteMessages.map { it.toEntity(topicName) }
+        val reactionEntities = remoteMessages.flatMap { it.reactions.toEntity(it.msgId, topicName) }
+        /* can be placed in the dao as a transaction,
+        but the signature of transaction method does not guarantee
+        that the entities actually have that topic name
+         */
+        messageDao.deleteMessages(topicName)
+        messageDao.insertMessages(messageEntities)
+        messageDao.insertReactions(reactionEntities)
+    }
+
+
+    fun getMessagesLocal(topicName: String): Single<List<MessageWithReactionsEntity>> {
+        return messageDao.getMessagesWithReactions(topicName)
+    }
+
+    fun getMessagesRemote(
+        streamName: String,
+        topicName: String,
+        anchorMessageId: Long,
+        numBefore: Int
+    ): Single<GetMessageResponse> {
         val narrow = Json.encodeToString(
             listOf(
-                Narrow("stream", streamName),
-                Narrow("topic", topicName)
+                Narrow(STREAM, streamName),
+                Narrow(TOPIC, topicName)
             )
         )
         return api.getMessages(
-            anchor = 10000000000000000,
-            numBefore = 50,
-            numAfter = 0,
+            anchor = anchorMessageId,
+            numBefore = numBefore,
             narrow = narrow,
             clientGravatar = false,
             applyMarkdown = false
-        )
+        ).subscribeOn(executionScheduler)
     }
 
-    fun uploadMoreMessages(
-        firstMessageId: Long,
-        streamName: String,
-        topicName: String
-    ): Single<List<MessageModel>> {
-        val narrow = Json.encodeToString(
-            listOf(
-                Narrow("stream", streamName),
-                Narrow("topic", topicName)
-            )
-        )
-        return api.getMessages(
-            firstMessageId,
-            20,
-            0,
-            narrow,
-            clientGravatar = false,
-            applyMarkdown = false
-        )
-            .map { messageResponse -> messageResponse.messages.toDomain() }
-    }
-
-    fun registerMessageQueue(
-        streamName: String,
-        topicName: String
-    ): Single<CreateQueueResponse> {
-        val narrow = (mapOf("stream" to streamName, "topic" to topicName))
-        val type = Json.encodeToString(listOf("message"))
-
-        return api.getQueue(type, narrow)
-            .subscribeOn(Schedulers.io())
+    fun registerMessageQueue(streamName: String, topicName: String): Single<CreateQueueResponse> {
+        val narrow = (mapOf(STREAM to streamName, TOPIC to topicName))
+        val type = Json.encodeToString(listOf(MESSAGE))
+        return api.getQueue(type, narrow).subscribeOn(executionScheduler)
     }
 
     fun getQueueMessages(
         currentMessageQueueId: String,
         lastId: Int
     ): Single<GetMessageEventsResponse> {
-        return api.getEventsFromQueue(currentMessageQueueId, lastId)
+        return api.getEventsFromQueue(currentMessageQueueId, lastId).subscribeOn(executionScheduler)
     }
 
-    fun registerReactionQueue(
-        streamName: String,
-        topicName: String
-    ): Single<CreateQueueResponse> {
-        val narrow = (mapOf("stream" to streamName, "topic" to topicName))
-        val type = Json.encodeToString(listOf("reaction"))
-
-        return api.getQueue(type, narrow)
-            .subscribeOn(Schedulers.io())
+    fun registerReactionQueue(streamName: String, topicName: String): Single<CreateQueueResponse> {
+        val narrow = (mapOf(STREAM to streamName, TOPIC to topicName))
+        val type = Json.encodeToString(listOf(REACTION))
+        return api.getQueue(type, narrow).subscribeOn(executionScheduler)
     }
 
     fun getQueueReactions(queueId: String, lastId: Int): Single<GetEmojiEventsResponse> {
-        return api.getEmojiEventsFromQueue(queueId, lastId)
+        return api.getEmojiEventsFromQueue(queueId, lastId).subscribeOn(executionScheduler)
     }
 
-    fun getMessage(messageId: Int) = api.getMessage(messageId)
-
-
-    fun onEmojiClick(emojiName: String, messageId: Int): Single<CreateReactionResponse> {
+    fun updateEmoji(emojiName: String, messageId: Int): Single<CreateReactionResponse> {
         return api.getMessage(id = messageId, applyMarkdown = true)
             .map { response ->
                 response.message.reactions.filter { reaction -> reaction.emojiName == emojiName }
                     .any { it.userId == Reactions.MY_USER_ID }
             }
-            .flatMap { isAlreadyClicked ->
-                updateEmoji(messageId, emojiName, isAlreadyClicked)
-            }
-            .subscribeOn(Schedulers.io())
+            .flatMap { isAlreadyClicked -> updateEmoji(messageId, emojiName, isAlreadyClicked) }
+            .subscribeOn(executionScheduler)
     }
 
     private fun updateEmoji(
@@ -130,54 +134,45 @@ class ChatRepository(private val messageDao: MessageDao, private val api: ZulipA
             to = streamName,
             topic = topicName,
             content = messageText
-        )
-            .subscribeOn(Schedulers.io())
+        ).subscribeOn(executionScheduler)
     }
 
-    private fun updateDb(
-        messagesInDb: List<MessageEntity>, newMessage: MessageResponse, topicName: String
-    ): Completable {
-        val msgExist = messagesInDb.any { it.msgId == newMessage.msgId }
-
-        return when {
-            msgExist -> {
-                Completable.fromCallable {
-                    messageDao.updateMessageReactions(
-                        newMessage.msgId,
-                        newMessage.reactions.toEntity(newMessage.msgId, topicName)
-                    )
+    fun saveMessage(newMessage: MessageResponse, topicName: String): Completable {
+        return messageDao.getMessages(topicName).flatMapCompletable { localMessages ->
+            val isMessageExists = localMessages.any { it.msgId == newMessage.msgId }
+            Completable.fromCallable {
+                when {
+                    isMessageExists -> {
+                        messageDao.updateMessageReactions(
+                            newMessage.msgId,
+                            newMessage.reactions.toEntity(newMessage.msgId, topicName)
+                        )
+                    }
+                    localMessages.size < 50 -> {
+                        messageDao.insertMessageWithReactions(
+                            newMessage.toEntity(topicName),
+                            newMessage.reactions.toEntity(newMessage.msgId, topicName)
+                        )
+                    }
+                    else -> {
+                        messageDao.deleteFirstAndAddNewMessage(
+                            localMessages.first().msgId,
+                            newMessage.toEntity(topicName),
+                            newMessage.reactions.toEntity(newMessage.msgId, topicName)
+                        )
+                    }
                 }
             }
-
-            messagesInDb.size < 50 -> {
-                Completable.fromCallable {
-                    messageDao.insertMessageWithReactions(
-                        newMessage.toEntity(topicName),
-                        newMessage.reactions.toEntity(newMessage.msgId, topicName)
-                    )
-                }
-            }
-
-            else -> {
-                Completable.fromCallable {
-                    messageDao.deleteFirstAndAddNewMessage(
-                        messagesInDb.first().msgId,
-                        newMessage.toEntity(topicName),
-                        newMessage.reactions.toEntity(newMessage.msgId, topicName)
-                    )
-                }
-            }
-        }
+        }.subscribeOn(executionScheduler)
     }
 
-
-    fun addToDatabase(
-        newMessage: MessageResponse, topicName: String
-    ): Completable {
-        return messageDao.getMessagesFromTopic(topicName)
-            .flatMapCompletable { messageEntityList ->
-                updateDb(messageEntityList, newMessage, topicName)
-            }
+    companion object {
+        const val STREAM = "stream"
+        const val TOPIC = "topic"
+        const val MESSAGE = "message"
+        const val REACTION = "reaction"
+        const val DEFAULT_MESSAGE_ANCHOR: Long = 10000000000000000
+        const val DEFAULT_NUM_BEFORE: Int = 50
+        const val MORE_NUM_BEFORE: Int = 20
     }
-
 }
